# [RFC] use dynamic dispatch for IdOrdMap Hash impl