--- a/crates/iddqd/tests/integration/id_ord_map.rs
+++ b/crates/iddqd/tests/integration/id_ord_map.rs
@@ -518,6 +518,20 @@ fn borrowed_item() {
     assert_eq!(map.get("foo").unwrap().key1, "foo");
     assert_eq!(map.get("bar").unwrap().key1, "bar");
 
+    // Check that we can mutably retrieve them.
+    {
+        let mut item1 = map.get_mut("foo").unwrap();
+        item1.key2 = b"foo2";
+
+        // Including reborrows.
+        {
+            let mut item1_reborrowed = item1.reborrow();
+            item1_reborrowed.key3 = Path::new("foo2");
+        }
+
+        item1.key2 = b"foo3";
+    }
+
     // Check that we can iterate over them.
     let keys: Vec<_> = map.iter().map(|item| item.key()).collect();
     assert_eq!(keys, vec!["bar", "foo"]);
@@ -532,7 +546,7 @@ fn borrowed_item() {
     static DEBUG_OUTPUT: &str = "{\"bar\": BorrowedItem { \
         key1: \"bar\", key2: [98, 97, 114], key3: \"bar\" }, \
         \"foo\": BorrowedItem { \
-        key1: \"foo\", key2: [102, 111, 111], key3: \"foo\" }}";
+        key1: \"foo\", key2: [102, 111, 111, 51], key3: \"foo2\" }}";
 
     assert_eq!(format!("{map:?}"), DEBUG_OUTPUT);
     assert_eq!(fmt_debug(&map), DEBUG_OUTPUT);
@@ -656,4 +670,57 @@ fn proptest_arbitrary_map(map: IdOrdMap<TestItem>) {
         assert_eq!(map.get(&item.key()), Some(item));
     }
     assert_eq!(count, len);
+}
+
+mod static_breakage {
+    use std::hash::Hash;
+
+    use super::*;
+
+    struct Item {
+        id: String,
+    }
+
+    impl IdOrdItem for Item {
+        type Key<'a> = Id<'a>;
+
+        fn key(&self) -> Self::Key<'_> {
+            Id(&self.id)
+        }
+
+        id_upcast!();
+    }
+
+    #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
+    struct Id<'a>(&'a str);
+
+    impl Hash for Id<'static> {
+        fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
+            self.0.hash(state);
+        }
+    }
+
+    #[test]
+    fn static_breakage() {
+        let mut map = IdOrdMap::new();
+        map.insert_overwrite(Item { id: "test".to_string() });
+
+        let map: &'static mut IdOrdMap<Item> = Box::leak(Box::new(map));
+
+        {
+            let item: id_ord_map::RefMut<'static, Item> =
+                map.get_mut(&Id("test")).unwrap();
+
+            let s = "test".to_owned();
+            let foo = Foo { item, s: &s };
+            // This drops the item, which calls the hash function above.
+            drop(foo);
+        }
+    }
+
+    #[expect(dead_code)]
+    struct Foo<'a> {
+        item: id_ord_map::RefMut<'a, Item>,
+        s: &'a str,
+    }
 }